backend/ai-agent/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
groq==0.9.0
pymongo==4.7.2
httpx==0.27.0
pydantic==2.6.4
pytest==8.2.0
pytest-asyncio==0.23.6
requests==2.31.0